split interfaces file into different categories
