+++ packages/data-model/src/classes/TypeFactory.ts
@@ -1,4 +1,5 @@
-import { iCharacterSheetData, iTraitData } from './../declarations/interfaces';
+import { iCharacterSheetData } from '../declarations/interfaces/character-sheet-interfaces';
+import { iTraitData } from './../declarations/interfaces';
 import { TraitMap, TraitName } from './../declarations/types';
 export default abstract class TypeFactory {
 	/*
