+++ packages/data-model/src/classes/log/AddLogEvent.ts
@@ -1,9 +1,8 @@
-import { iAddLogEvent } from './../../declarations/interfaces';
-import { iBaseLogEventProps, iLogEvent, iNewValue } from '../../declarations/interfaces';
+import { iNewValue } from '../../declarations/interfaces/general-interfaces';
+import { iBaseLogEventProps, iAddLogEvent } from '../../declarations/interfaces/log-interfaces';
 import BaseLogEvent from './BaseLogEvent';
 
 interface iProps<T> extends iBaseLogEventProps, iNewValue<T> {}
-
 
 // ? do these need to implement iNewValue? whenthese are put into a collection, the additional iNewValue, iOldValue etc details are lost so is there a point? the differences shouldjust be on the specific classes and they all extend iBaseLogEvent? Actually properties can still be accessed using instanceof checks?
 export default class AddLogEvent<T> extends BaseLogEvent<T> implements iAddLogEvent<T> {
@@ -13,7 +12,7 @@ export default class AddLogEvent<T> extends BaseLogEvent<T> implements iAddLogEv
 		super({ operation: 'ADD', property, note: description });
 		this.newValue = newValue;
 	}
-  describe(): string {
-    return `${this.property} added with inital value of ${this.newValue}`
-  }
+	describe(): string {
+		return `${this.property} added with inital value of ${this.newValue}`;
+	}
 }
