+++ packages/data-model/src/classes/log/DeleteLogEvent.ts
@@ -1,4 +1,6 @@
-import { iBaseLogEventProps, iLogEvent, iOldValue } from '../../declarations/interfaces';
+import {  } from '../../declarations/interfaces';
+import { iOldValue } from '../../declarations/interfaces/general-interfaces';
+import { iBaseLogEventProps, iLogEvent } from '../../declarations/interfaces/log-interfaces';
 import BaseLogEvent from './BaseLogEvent';
 
 interface iProps<T> extends iBaseLogEventProps, iOldValue<T> {}
