+++ packages/data-model/src/classes/traits/BaseTrait.ts
@@ -1,13 +1,7 @@
+import { iLogEvent } from './../../declarations/interfaces/log-interfaces';
 import { iBaseTrait } from './../../declarations/interfaces';
 import { TraitData, TraitName, TraitValue } from '../../declarations/types';
-import {
-	iBaseTraitProps,
-	iCharacterSheet,
-	iLogEvent,
-	iLogger,
-	iToJson,
-	iTraitData,
-} from '../../declarations/interfaces';
+import { iBaseTraitProps, iTraitData } from '../../declarations/interfaces';
 import LogCollection from '../log/LogCollection';
 import UpdateLogEvent from '../log/UpdateLogEvent';
